Remove the experimental offscreen item modules
