+++ app/src/sandbox/combobox-select-selected-value-presentation/test.ts
@@ -1,0 +1,12 @@
+import { click, q, type } from "@ariakit/test";
+import { expect, test } from "vitest";
+
+// https://github.com/ariakit/ariakit/pull/6832
+test("activates a far item reached through typeahead", async () => {
+  const select = q.combobox("Selected fruit");
+  await click(select);
+  await type("ly");
+
+  expect(q.option("Lychee")).toHaveAttribute("data-active-item");
+  expect(select).toHaveFocus();
+});
